+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Bugs fixed
+
+* Fix bug concerning table and separator alignment of multi-line hash with multiple keys on the same line.
+
 ## 0.14.0 (07/10/2013)
 
 ### New features